projects/dthostmon/requirements.txt
# Core Dependencies for dthostmon Phase 1
# Last Updated: 11/14/2025 12:00:00 PM CDT

# SSH and Remote Access
paramiko>=3.4.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# API Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0

# AI/LLM Integration
openai>=1.3.5
requests>=2.31.0

# Email
email-validator>=2.1.0

# Utilities
python-dateutil>=2.8.2
